Made the filter projection a factor n faster
The current filter projection has a O(n*n) time complexity, when the
filter mapping change for all items (think searching). For each item
update the projection updates the indexes to the end of the array.

I'm assuming that the mapping function will be cheap in the case of
filtering, what it would usually would be. When an inclusion change, it
updates the mapping to the end of the array immidiately. That means the
updates comming after the first update has to do not work.